+++ packages/core/src/constant.ts
@@ -32,10 +32,14 @@ export const SOLID_CLIENT_AUTHN_KEY_PREFIX = "solidClientAuthn:";
 export const PREFERRED_SIGNING_ALG = ["ES256", "RS256"];
 
 export const EVENTS = {
+  // Note that an `error` events MUST be listened to: https://nodejs.org/dist/latest-v16.x/docs/api/events.html#error-events.
+  ERROR: "error",
+  LOGIN: "login",
+  LOGOUT: "logout",
   NEW_REFRESH_TOKEN: "newRefreshToken",
-  ERROR: "onError",
   SESSION_EXPIRED: "sessionExpired",
   SESSION_EXTENDED: "sessionExtended",
+  SESSION_RESTORED: "sessionRestore",
   TIMEOUT_SET: "timeoutSet",
 };
 /**